requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bidict==0.23.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.8.0
granian==2.7.2
greenlet==3.3.1
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.13
langchain-groq==1.1.2
langchain-text-splitters==1.1.0
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.6
langsmith==0.7.3
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.2
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
platformdirs==4.9.2
propcache==0.4.1
psutil==7.2.2
pydantic==2.12.5
pydantic-settings==2.13.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-engineio==4.13.1
python-multipart==0.0.22
python-socketio==5.16.1
PyYAML==6.0.3
redis==7.2.1
reflex==0.8.27
reflex-hosting-cli==0.1.62
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.46
sqlmodel==0.0.37
starlette==0.52.1
tenacity==9.1.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
watchfiles==1.1.1
wrapt==2.1.1
wsproto==1.3.2
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
